ref(pkg/catalog): clean up some unused funcs
part of #2397
